refactor: create pages folder and move about, main, not-found
