--- src/components/about/about.css
@@ -1,10 +1,0 @@
-.about {
-    margin-top: 2rem;
-}
-
-.about-content {
-    display: flex;
-    margin: 0 auto;
-    max-width: 860px;
-    width: 100%;
-}
