+++ src/pages/not-found/NotFound.tsx
@@ -4,7 +4,7 @@ import { NavLink } from 'react-router-dom';
 import astronaut from '../../assets/astronaut.svg';
 import planet from '../../assets/planet.svg';
 
-import './not-found.css';
+import './Not-found.css';
 
 class NotFound extends Component {
   render(): ReactNode {
